+++ ir/pointer.cpp
@@ -327,7 +327,7 @@ expr Pointer::isBlockAligned(uint64_t align, bool exact) const {
   return exact ? blk_align == bits : blk_align.uge(bits);
 }
 
-expr Pointer::isAligned(unsigned align) {
+expr Pointer::isAligned(uint64_t align) {
   if (align == 1)
     return true;
 
@@ -360,7 +360,7 @@ expr Pointer::isAligned(unsigned align) {
 static pair<expr, expr> is_dereferenceable(Pointer &p,
                                            const expr &bytes_off,
                                            const expr &bytes,
-                                           unsigned align, bool iswrite) {
+                                           uint64_t align, bool iswrite) {
   expr block_sz = p.blockSizeOffsetT();
   expr offset = p.getOffset();
 
@@ -387,7 +387,7 @@ static pair<expr, expr> is_dereferenceable(Pointer &p,
 }
 
 // When bytes is 0, pointer is always derefenceable
-AndExpr Pointer::isDereferenceable(const expr &bytes0, unsigned align,
+AndExpr Pointer::isDereferenceable(const expr &bytes0, uint64_t align,
                                    bool iswrite) {
   expr bytes_off = bytes0.zextOrTrunc(bits_for_offset);
   expr bytes = bytes0.zextOrTrunc(bits_size_t);
@@ -425,7 +425,7 @@ AndExpr Pointer::isDereferenceable(const expr &bytes0, unsigned align,
   return exprs;
 }
 
-AndExpr Pointer::isDereferenceable(uint64_t bytes, unsigned align,
+AndExpr Pointer::isDereferenceable(uint64_t bytes, uint64_t align,
                                    bool iswrite) {
   return isDereferenceable(expr::mkUInt(bytes, bits_size_t), align, iswrite);
 }
